+++ checker/src/main/java/org/checkerframework/checker/nullness/NullnessAnnotatedTypeFactory.java
@@ -371,11 +371,6 @@ public class NullnessAnnotatedTypeFactory
         NULLABLE_ALIASES.forEach(annotation -> addAliasedTypeAnnotation(annotation, NULLABLE));
         POLYNULL_ALIASES.forEach(annotation -> addAliasedTypeAnnotation(annotation, POLYNULL));
 
-        // Add compatible annotations from alias file
-        addAliasedTypeAnnotation(NonNull.class);
-        addAliasedTypeAnnotation(Nullable.class);
-        addAliasedTypeAnnotation(PolyNull.class);
-
         // Add compatibility annotations:
         addAliasedTypeAnnotation(
                 "org.checkerframework.checker.nullness.compatqual.PolyNullDecl", POLYNULL);