+++ checker/src/main/java/org/checkerframework/checker/signedness/SignednessAnnotatedTypeFactory.java
@@ -111,9 +111,6 @@ public class SignednessAnnotatedTypeFactory extends BaseAnnotatedTypeFactory {
 
         addAliasedTypeAnnotation("jdk.jfr.Unsigned", UNSIGNED);
 
-        // Add compatible annotations from alias file
-        addAliasedTypeAnnotation(Unsigned.class);
-
         postInit();
     }