+++ framework/src/main/java/org/checkerframework/framework/util/PurityAnnotatedTypeFactory.java
@@ -2,7 +2,6 @@ package org.checkerframework.framework.util;
 
 import org.checkerframework.common.basetype.BaseAnnotatedTypeFactory;
 import org.checkerframework.common.basetype.BaseTypeChecker;
-import org.checkerframework.dataflow.qual.Pure;
 import org.checkerframework.framework.qual.PurityUnqualified;
 
 import java.lang.annotation.Annotation;
@@ -15,10 +14,6 @@ public class PurityAnnotatedTypeFactory extends BaseAnnotatedTypeFactory {
 
     public PurityAnnotatedTypeFactory(BaseTypeChecker checker) {
         super(checker);
-
-        // Add compatible annotations from alias file
-        addAliasedTypeAnnotation(Pure.class);
-
         this.postInit();
     }