+++ docs/CHANGELOG.md
@@ -2,6 +2,8 @@ Version 3.32.0-eisop2 (March ?, 2023)
 -------------------------------------
 
 **User-visible changes:**
+The new command-line argument `-Aaliases={folders}` enables user to define custom annotations
+to be treat as Checker Framework annotations.
 
 **Implementation details:**